LabFiles/beanbotics-troubleshooting/requirements.txt
# BeanBotics Ticketing System Requirements
fast-agent-mcp==0.3.15
fastmcp==2.12.4
flask-cors==4.0.0
flask-openapi3[redoc]==4.3.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask==3.0.0
python-socketio[client]==5.11.0
requests==2.31.0